+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added an `ElementContent` variant whose `measureFunction` takes in both a `SizeConstraint` and an `Environment`.
+
 ### Removed
 
 ### Changed